refactor: rename `cache.ts` to `lib.ts`
